feat(grid): improve value selector component
